lib: Downgrade fp_err() to be non-fatal
Similarly to b1ac865abdc017bb824aab452da37e026d3a9a4c, downgrade
fp_err() to be non-fatal. A number of drivers would spit out an error
when encountering this call, but not crash, carry on and most of the
time recover.

Make sure we don't assert in those cases.